fix: pagination `Links` header parsing bug
The PandaScore servers seem to accept headers in the form of
`page[number]=x` when requesting. However, this seems to cause bugs when
reading the `Links` header as the server will have both `page[number]`
and `page` in the returned link.

The solution here is to not use `page[number]` at all.